+++ Runtime/Cartridges/Cartridge.cs
@@ -5,6 +5,7 @@ namespace Pixygon.Micro {
     public class Cartridge : ScriptableObject {
         public string _title;
         public string _description;
+        public string _version;
         public GameObject _gamePrefab;
         public Texture2D _cartridgeImage;
         public Color _cartridgeColor;